running code

+++ App.java
@@ -3,24 +3,26 @@ import java.util.*;
 public class App {
 
     public static Vector<VisitorAccount> accounts = new Vector<VisitorAccount>();
-    public static Vector<VisitorAccount> admins = new Vector<VisitorAccount>();
+    public static VisitorAccount admins = new VisitorAccount();
     public static Vector<Course> coursesVector = new Vector<Course>();
     
 
     public static void main(String args[]) 
     {
-
+        Course course;
+        coursesVector.add(course = new Course("math 0", "Math", "MA12", "", 100.0, "material"));
+        admins.Email = "admin";
+        admins.Pass = "admin";
         String id;
         VisitorAccount user = new VisitorAccount();
         Scanner scanner = new Scanner(System.in);
         Administration admin = new Administration();
         boolean loopBreak = false;
         int choice;
-        //admin.setCourse();
         for (;;) 
         {
             System.out.println("         Welcome to Training Advisor");
-            System.out.println("Press 1 to Signup as a user, 2 to signup as an administrator and 3 to exit");
+            System.out.println("Press 1 to Signup as a user, 2 to Login as an administrator and 3 to exit");
             choice = scanner.nextInt();
 
             switch (choice) {
@@ -38,7 +40,7 @@ public class App {
                         while (loopBreak == false) {
                             System.out.println("      Main menu");
                             System.out.println(
-                                    "View my courses (Press 1) \nSearch course (Press 2) \nEnroll Course (Press 3) \nEdit Profile (Press 4) \nWithdraw from a Course (Press 5) \nView Profile (Press 6) \nSignup from another account (Press 7) \nLogin from another account (Press 8) \n get certificate (Press 9) \nExit to login page (Press 0) ");
+                                    "View my courses (Press 1) \nSearch course (Press 2) \nEnroll Course (Press 3) \nEdit Profile (Press 4) \nWithdraw from a Course (Press 5) \nView Profile (Press 6) \nSignup from another account (Press 7) \nLogin from another account (Press 8) \n get certificate (Press 9)\nSolve quizes and exams (Press 10) \nExit to login page (Press 0) ");
                             choice = scanner.nextInt();
                             switch (choice) {
                                 case 0:
@@ -52,14 +54,36 @@ public class App {
                                     else {
                                         for (int i = 0; i < accounts.get(VisitorAccount.accountIndex).myCourses
                                                 .size(); i++) {
-                                            System.out
-                                                    .println(accounts.get(VisitorAccount.accountIndex).myCourses.get(i)
+                                            System.out.println("Course Name: "
+                                                    + accounts.get(VisitorAccount.accountIndex).myCourses.get(i)
                                                             .getCourseName());
+                                            System.out.println(
+                                                    "Course ID: " + accounts.get(VisitorAccount.accountIndex).myCourses
+                                                            .get(i).getCourseID());
+                                        }
+                                        System.out.println(
+                                                "If you want to track progress of a course press one, else press any other key");
+                                        choice = scanner.nextInt();
+                                        if (choice == 1) {
+                                            System.out.println("Enter course ID");
+                                            scanner.nextLine();
+                                            String iDString = scanner.nextLine();
+                                            for (int i = 0; i < accounts.get(VisitorAccount.accountIndex).myCourses
+                                                    .size(); i++) 
+                                            {
+                                                if (accounts.get(VisitorAccount.accountIndex).myCourses
+                                                        .get(VisitorAccount.accountIndex).getCourseID()
+                                                        .equals(iDString)) 
+                                                {
+                                                    accounts.get(VisitorAccount.accountIndex).myCourses.get(VisitorAccount.accountIndex).progressTracking();
+                                                }
+                                            }
+
                                         }
                                     }
                                     break;
                                 case 2:
-                                    CourseSearch courseSearch = new CourseSearch();
+                                    Course courseSearch = new Course();
                                     courseSearch.SearchCourse();
 
                                     break;
@@ -104,6 +128,49 @@ public class App {
                                     CertificateIssuance Certificate = new CertificateIssuance();
                                     Certificate.issueCertificate(id);
                                     break;
+                                case 10:
+                                    System.out.println("Press 1 to solve the exam and 2 to solve the quiz");
+                                    choice = scanner.nextInt();
+                                    switch (choice) {
+                                        case 1:
+                                            System.out.println("Enter course ID");
+                                            scanner.nextLine();
+                                            id = scanner.nextLine();
+                                            id.toUpperCase();
+
+                                            for (int i = 0; i < accounts.get(VisitorAccount.accountIndex).myCourses
+                                                    .size(); i++) {
+                                                if (accounts.get(VisitorAccount.accountIndex).myCourses
+                                                        .get(VisitorAccount.accountIndex).getCourseID().equals(id)) {
+                                                    accounts.get(VisitorAccount.accountIndex).myCourses
+                                                            .get(VisitorAccount.accountIndex).exam
+                                                            .solveExam(
+                                                                    accounts.get(VisitorAccount.accountIndex).myCourses
+                                                                            .get(VisitorAccount.accountIndex)
+                                                                            .getCourseScore());
+                                                }
+                                            }
+                                            break;
+                                        case 2:
+                                            System.out.println("Enter course ID");
+                                            scanner.nextLine();
+                                            id = scanner.nextLine();
+                                            id.toUpperCase();
+
+                                            for (int i = 0; i < accounts.get(VisitorAccount.accountIndex).myCourses
+                                                    .size(); i++) {
+                                                if (accounts.get(VisitorAccount.accountIndex).myCourses
+                                                        .get(i).getCourseID().equals(id)) 
+                                                        {
+                                                    accounts.get(VisitorAccount.accountIndex).myCourses.get(i).exam
+                                                            .solveQuiz(
+                                                                    accounts.get(VisitorAccount.accountIndex).myCourses.get(i)
+                                                                            .getCourseScore());
+                                                }
+                                            }
+                                            break;
+                                    }
+                                    break;
                                 default:
                                     System.out.println("Invalid input");
                             }
@@ -112,32 +179,41 @@ public class App {
                     break;
                 case 2:
                     loopBreak = false;
-                    admins.add(user.Signup());
-                    System.out.println("Signed up successfully");
-                    while (loopBreak == false) {
-                        System.out.println(
-                                "Set course (Press 1) \nEdit course (Press 2) \nDelete Course(Press 3) \nExit to main menu (Press 0)");
-                        choice = scanner.nextInt();
-                        switch (choice) {
-                            case 1:
-                                admin.setCourse();
-                                break;
-                            case 2:
-                                System.out.println("Enter course ID");
-                                id = scanner.nextLine();
-                                id.toUpperCase();
-                                admin.DefiningCourse(id);
-                                break;
-                            case 3:
-                                System.out.println("Enter course ID");
-                                id = scanner.nextLine();
-                                id.toUpperCase();
-                                admin.deleteCourse(id);
-                            case 0:
-                                loopBreak = true;
-                                break;
-                            default:
-                                System.out.println("Invalid choice");
+                    String inputUser, inputPass;
+                    System.out.print("Enter the username: ");
+                    scanner.nextLine();
+                    inputUser = scanner.nextLine();
+                    System.out.print("Enter the password: ");
+                    inputPass = scanner.nextLine();
+                    if (inputPass.equals(admins.Pass) && inputUser.equals(admins.Email))
+                    {
+                        System.out.println("Logged up successfully");
+                        while (loopBreak == false) {
+                            System.out.println(
+                                    "Set course (Press 1) \nEdit course (Press 2) \nDelete Course(Press 3) \nExit to main menu (Press 0)");
+                            choice = scanner.nextInt();
+                            scanner.nextLine();
+                            switch (choice) {
+                                case 1:
+                                    admin.setCourse();
+                                    break;
+                                case 2:
+                                    System.out.println("Enter course ID");
+                                    id = scanner.nextLine();
+                                    id.toUpperCase();
+                                    admin.DefiningCourse(id);
+                                    break;
+                                case 3:
+                                    System.out.println("Enter course ID");
+                                    id = scanner.nextLine();
+                                    id.toUpperCase();
+                                    admin.deleteCourse(id);
+                                case 0:
+                                    loopBreak = true;
+                                    break;
+                                default:
+                                    System.out.println("Invalid choice");
+                            }
                         }
                     }
                     break;
